+++ Windows/include/MF/StringSafePlaceHolder.hpp
